% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{new_foo}
\alias{new_foo}
\title{Constructor}
\usage{
new_foo(x)
}
\arguments{
\item{x}{numeric}
}
\value{
object of class foo
}
\description{
\code{new_foo} is a constructor for class foo
}
\examples{
nf <- new_foo(3)
}
