% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg1r.R
\docType{package}
\name{pkg1r}
\alias{pkg1r}
\alias{pkg1r-package}
\title{pkg1r: A package for doing nothing.}
\description{
The pkgr1 package provides three categories of important functions:
test, test1 and test2.
}
\section{pkg1r functions}{

The pkg1r functions ...
}

