% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{showoff.foo}
\alias{showoff.foo}
\title{Big time extrovert for foo}
\usage{
\method{showoff}{foo}(x)
}
\arguments{
\item{x}{numeric}
}
\value{
show off number for foo
}
\description{
\code{showoff} is a method for class foo of generic
}
\examples{
nf <- new_foo(3)
showoff(nf)
}
