% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test function}
\usage{
test(x)
}
\arguments{
\item{x}{numeric}
}
\value{
the square of \code{x}
}
\description{
\code{test} actually return the square of its input
}
\examples{
test(3L)
test(3.5)

\dontrun {
test('a')
}
}
